credit card authorization: Isolate the card instantiation from hash
The card build was in the authorize method. It moves the behavior to a new method called `generateCard` and includes some integration tests to garantee that it works well mocking some SDK's methods. To do that it was necessary to refactor some lines like remove `Mage` static calls and create a new property for SDK (composition).