creditcard: check if pagarme exception is parseable json
As you can se on https://github.com/pagarme/pagarme-php/blob/V3/lib/PagarMeException.php#L12 the exception message could be just a string, so an invalid json.
If the message is a string the next steps will broke the application. This includes a verification to check if the exceptions message can be parsed to json