Improve postback requests handling (#335)
* postback: create basic structure to handler postbacks

Define an abstraction for postback handlers and an especialization to
handler a "paid postback". Also create an exception to expose problems
during the handling and a factory to instantiate the correct handler
based on desired status

* postback: use "paid postback" with new handler structure

Changes the Postback model to use the postback handler structure created previously

* postback: handle paid postback from processing old status

If a given transaction is created async its order is setted to
payment_review. So, the paid postback should update
the order to processing state and then invoice it.

* postback: set required values on invoice

* postback: create the authorize postback handler

Creates a authorize postback handler responsible to move orders to
processing (magento state) and then wait for a Paid postback that will
be handled by the paid handler created previously

* postback: create refund postback handler

* postback: create a refused postback handler

* review: apply required changes

* module version: bump module version to v3.13.1