Implement a custom orders state/status workflow (#337)
* postback: set auth_only order desired state to pending_payment

Since the `processing` state means that the transaction related to an
order is paid. Authorized only orders' state should be
`pending_payment`. This change will executed during a postback handling
request

* config: create consts on payment action config class

The previous version was using literals for options. This create consts
to be used on the config array also, allow to use that consts externally

* xml config: set the default order state to pending_payment

Every order that will be created should have the `pending_payment`
status and it will be handled programmatically by the payment method
model

* credit card: handle auth_only/capture workflow state

Give up the magento's default state workflow to be able to set orders as
`pending_payment` created by _auth only_ or _async_ behaviors

* dependencies: update pagarme-php sdk

Update projects dependency to use the latest pagarme's sdk version

* credit card: handle status sent by API

This handle the status informed by Pagar.me's API and set the order
state accordingly.

* credit card: create invoice for paid transactions

Create an invoice for paid transactions and refactor the
`handlePaymentStatus` method.

* credit card: deprecate order observer

This observer was necessary because the module was using magento's
default status workflow. Since the module implements its own workflow
this observer isn't necessary anymore.

* locale: translate and adjust messages on order's history

Since new messages will be created to be displayed on order's history
comments. This create message entries on locale files to translate that
messages to pt_br.

* tests: implement e2e tests to validate order status

The order status should be `pending_payment` or `processing` based on
module's configuration. This tests guarantee that the expected order
status is working well

* module version: bump module version to v3.14.0

* tests: remove out of date unit test

Since the method contract/structure is changed. This test is out of
date. Also, this test, now is using some magento implementations that is
dificult to mock, so its visibility is changed to `private`

* code review: apply suggested changes