Adds payment information on buyer's admin area (#318)
* credit card: Adds payment information on buyer's admin area

The buyer can't know the payment information on its administration area,
so this commit adds this information to give a better visibility to the
buyer.

* credit card: remove title from payment details block

Since this template is displayed on customer panel, print and email
templates and the order was purchased using transparent checkout doesn't
makes sense display the gateway name used.

* boleto: create a template for payment details

Create a payment details to be displayed on customer panel, email and
print templates

* module version: bump to v3.14.3

* boleto: fix payment method template

The payment_method template is the same on admin details, emails and
checkout so, this creates a new template to be showed during checkout
process which won't has a pagarme transaction created yet. This new
template will show the payment method title only.

* credit card: refactor payment_method template during checkout

Previously a fix was implemented to show a different template during
checkout when the payment method is boleto. This is a refactor to use
the same structure.

* code review: apply requested changes