fix interest rate calculation (#342)
* code standard: define and ignore rules

Some rules required during on code review wasn't check by the code
sniffer. Also some rules should be ignored. It adds and ignore those
rules

* credit_card: remove PagarMe_Core_Block_Info_Trait transaction variable

it is already declared in PagarMe_CreditCard_Block_Info_Creditcard

* credit_card: create installments calculator

it will improve the installments calculation, and the retrieve of the
order with installments total.

* credit_card: fix checkout installments select box exibition

Now the installments select box retrieve the value of the shipping +
cart items total, considering it discounts.

* credit_card: fix order review grand total

Changed the review grand total to include the interest rate.
Now the interest rate is calculated based on shipping amount, cart
items total and discount rules.

* coding_standards: adapt files to follow phpcs rules

* core: fix interest amount value saved on transaction

With the previous changes, the grand amount has been changed,
considering the interest amount in it.
To fix it, we got the shipping amount + products total amount and
subtract it from the order total (with the interest amount), to retrieve the
real interest amount.

* module version: bump to v3.14.4