Allow order creation from admin panel (#367)
* #358 fixes order creation in backend using Pagarme_Boleto

* tests: create scenario to validate admin order creation

Describes, step-by-step the workflow to create orders from admin
choosing boleto as payment method

* tests: refactor contexts classes

There's a lot of repeated code distributed on context classes. It
creates a new classe to be used as base. It'll centralize all the steps
and usefull methods that can be shared on different suites.

* tests: implement e2e tests for boleto purchases made from admin

Creates e2e tests to validate that a store admin user can create orders
from admin using Pagar.me's Boleto payment method. To do that, creates a
new context class to isolate the code responsible for this.

* tests: implement workaround to set street lines config to 4

I don't like this implementation but, since that's the only way I found
to finish this task I'll use it. It sets the street lines config to 4
lines to allow admin/customers to buy on platform using Pagar.me's
module. This workaround sets the value programmatically but, to this
change be saved properly a press button on the admin page was needed.