Migrate to Circle-Ci (#376)
* ci: migrate from travis to circle-ci

It creates a new config file to allow us to use Circle-ci to build and
test the module

* script: avoid to check forgotten keys on vendor dir

* make: create a makefile to improve development workflow

It's difficult and annoying to remember all commands to run the project.
It creates a make file that improve the development experience.

* circle-ci: adjust config to use makefile commands

* circle-ci: add before-deps parameter

The `postback` test suite is depends `configure` suite. This commit add
a new parameter that allow to run this dependency suite only for
postback test suite

* circle-ci: create job to install composer dependencies

It'll install project dependencies after checkou-code and before all the
tests tha will use those dependencies

* tests: increase delay to check url

There's a step that check if, when a customer receives refused message,
the page continues on checkout page.

* circleci: create parameterized job for unit tests

Unit tests should be executed against/considering magento version. It
creates adds parameters to choose the magento version that will be
used. Also, to improve the workflow, set e2e suites as dependens to each
unit job in the workflow

* make: create command to run phpcs

* doc: adding instructions to setup project

* script: add shebang