checkout: add card animation
Add pagarme/card component to magento checkout.

First, we need to know that the component provides the input masks, in
the following format.
- credit card number: depends on the brand, but the common is `**** ****
**** ****`
- expiration date: MM / YY

Because of the component masks, we had to remove the `validate-digits` validator
from credit card number and expiration date inputs

We added the `validate-alphanum-with-spaces` form validator to credit
card owner, and credit card number to guarantee the data integrity, and
because the component mask just allow numbers on credit card number
input.